ENG-1814 Use 4 bytes for int32, int16 and int8 instead of 8 bytes.
Summary:
We currently use fixed 8 bytes for all types of ints (int8, int16, int32 and int64). As a
result, for smaller int types we are wasting a lot of space in memory and on disk. With this diff,
we ensure int8, int16 and int32 use 4 bytes. I'll follow up this diff with another diff to change
float to 4 bytes.

Note that this is a format change.

Test Plan: Enhanced unit tests for int32, jenkins

Reviewers: mikhail, akashnil, hector

Reviewed By: hector

Subscribers: hector, ybase, kannan, karthik

Differential Revision: https://phabricator.dev.yugabyte.com/D2503

+++ src/yb/docdb/doc_rowwise_iterator.cc
@@ -214,14 +214,14 @@ CHECKED_STATUS PrimitiveValueToKudu(const Schema& projection,
     }
     case DataType::INT32: {
       // TODO: update these casts when all data types are supported in docdb
-      RETURN_NOT_OK(ExpectValueType(ValueType::kInt64, col_schema, value));
-      *(reinterpret_cast<int32_t*>(dest_ptr)) = static_cast<int32_t>(value.GetInt64());
+      RETURN_NOT_OK(ExpectValueType(ValueType::kInt32, col_schema, value));
+      *(reinterpret_cast<int32_t*>(dest_ptr)) = static_cast<int32_t>(value.GetInt32());
       break;
     }
     case DataType::INT8: {
       // TODO: update these casts when all data types are supported in docdb
-      RETURN_NOT_OK(ExpectValueType(ValueType::kInt64, col_schema, value));
-      *(reinterpret_cast<int8_t*>(dest_ptr)) = static_cast<int8_t>(value.GetInt64());
+      RETURN_NOT_OK(ExpectValueType(ValueType::kInt32, col_schema, value));
+      *(reinterpret_cast<int8_t*>(dest_ptr)) = static_cast<int8_t>(value.GetInt32());
       break;
     }
     case DataType::BOOL: {
